% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_cmbn_exptsmry}
\alias{copy_cmbn_exptsmry}
\title{Copy an \code{expt_smry...} file to \code{dat_dir}}
\usage{
copy_cmbn_exptsmry(dat_dir)
}
\description{
Copy an \code{expt_smry...} file to \code{dat_dir}
}
