% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_cmbn_phospho_mascot}
\alias{copy_cmbn_phospho_mascot}
\title{Copy Mascot phospho and global \code{.csv}}
\usage{
copy_cmbn_phospho_mascot(dat_dir)
}
\description{
Copy Mascot phospho and global \code{.csv}
}
