% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_crap}
\alias{copy_crap}
\title{Copy \code{crpa.fasta} to \code{db_path}}
\usage{
copy_crap(db_path = "~/proteoQ/dbs/fasta/crap")
}
\arguments{
\item{db_path}{Character string; the local path for database(s).}
}
\description{
Copy \code{crpa.fasta} to \code{db_path}
}
