% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_exptsmry_cmbn}
\alias{copy_exptsmry_cmbn}
\title{Copy a metadata file \code{expt_smry.xlsx} file to \code{dat_dir}}
\usage{
copy_exptsmry_cmbn(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy a metadata file \code{expt_smry.xlsx} file to \code{dat_dir}
}
