% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_fasta}
\alias{copy_fasta}
\title{Copy \code{fasta} files}
\usage{
copy_fasta(fasta_dir = "~\\\\proteoQ\\\\dbs\\\\fasta\\\\refseq",
  from = "refseq_hs_2013_07.fasta", to = "refseq_hs_2013_07.fasta")
}
\description{
\code{copy_fasta} copies \code{fasta} files to a target directory.
}
