% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_frac}
\alias{copy_frac}
\title{Copy \code{frac_smry.xlsx}}
\usage{
copy_frac(dat_dir, filename = "frac_smry.xlsx")
}
\description{
\code{copy_frac} copies the \code{frac_smry.xlsx} to a target directory.
}
