% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_fracsmry_gtmt}
\alias{copy_fracsmry_gtmt}
\title{Copy a metadata file \code{frac_smry.xlsx} to \code{dat_dir}}
\usage{
copy_fracsmry_gtmt(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy a metadata file \code{frac_smry.xlsx} to \code{dat_dir}
}
