% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_global_fracsmry_lfq}
\alias{copy_global_fracsmry_lfq}
\title{Copy a metadata file \code{frac_smry.xlsx} to \code{dat_dir}}
\usage{
copy_global_fracsmry_lfq(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Copy a metadata file \code{frac_smry.xlsx} to \code{dat_dir}
}
