% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_global_maxquant_lfq}
\alias{copy_global_maxquant_lfq}
\title{Copy MaxQuant LFQ global \code{.txt}}
\usage{
copy_global_maxquant_lfq(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Copy MaxQuant LFQ global \code{.txt}
}
