% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_global_sm}
\alias{copy_global_sm}
\title{Copy Spectrum Mill global \code{.ssv}}
\usage{
copy_global_sm(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Copy Spectrum Mill global \code{.ssv}
}
