% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_globalCPTAC}
\alias{copy_globalCPTAC}
\title{Copy Mascot \code{.csv} files}
\usage{
copy_globalCPTAC(dat_dir)
}
\description{
\code{copy_globalCPTAC} copies the \code{.csv} files from Mascot searches to
a target directory.
}
