% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_mascot_ptmt}
\alias{copy_mascot_ptmt}
\title{Copy Mascot phospho TMT \code{.csv}}
\usage{
copy_mascot_ptmt(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy Mascot phospho TMT \code{.csv}
}
