% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_maxquant_gtmt}
\alias{copy_maxquant_gtmt}
\title{Copy MaxQuant global TMT \code{.txt}}
\usage{
copy_maxquant_gtmt(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy MaxQuant global TMT \code{.txt}
}
