% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_maxquant_plfq}
\alias{copy_maxquant_plfq}
\title{Copy MaxQuant phospho LFQ \code{.txt}}
\usage{
copy_maxquant_plfq(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy MaxQuant phospho LFQ \code{.txt}
}
