% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_maxquant_ptmt}
\alias{copy_maxquant_ptmt}
\title{Copy MaxQuant phospho TMT \code{.txt}}
\usage{
copy_maxquant_ptmt(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy MaxQuant phospho TMT \code{.txt}
}
