% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_mgf}
\alias{copy_mgf}
\title{Copy \code{.mgf} files}
\usage{
copy_mgf(dat_dir = "~/proteoQ/examples/mgfs", filelist, mapped = NULL)
}
\arguments{
\item{dat_dir}{A character string to the working directory.}

\item{filelist}{A list of files}

\item{mapped}{A mapping suffix.
\@examples 
\donttest{
copy_mgf("~/proteoQ/examples/mgfs", "msconv_bi1_gl_partial)
copy_mgf("~/proteoQ/examples/mgfs", "pd_bi1_imac_partial)
}}
}
\description{
\code{copy_mgf} copies \code{.mgf} files to a target directory.
}
