% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_mq_txt}
\alias{copy_mq_txt}
\title{Copy MaxQuant \code{.txt} files}
\usage{
copy_mq_txt(dat_dir, filelist)
}
\description{
\code{copy_mq_txt} copies the MaxQuant outputs of \code{.txt} files to a
target directory.
}
