% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_msconv_mgf}
\alias{copy_msconv_mgf}
\title{Copy MSConvert MGF files.}
\usage{
copy_msconv_mgf(dat_dir = "~/proteoQ/examples/mgfs")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy MSConvert MGF files.
}
