% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_msfragger_gtmt}
\alias{copy_msfragger_gtmt}
\title{Copy MSFragger global TMT \code{.tsv}}
\usage{
copy_msfragger_gtmt(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy MSFragger global TMT \code{.tsv}
}
