% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_msfragger_plfq}
\alias{copy_msfragger_plfq}
\title{Copy MSFragger phospho LFQ \code{.tsv}}
\usage{
copy_msfragger_plfq(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy MSFragger phospho LFQ \code{.tsv}
}
