% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_pd_mgf}
\alias{copy_pd_mgf}
\title{Copy Proteome Discoverer MGF files.}
\usage{
copy_pd_mgf(dat_dir = "~/proteoQ/examples/mgfs")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy Proteome Discoverer MGF files.
}
