% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_phospho_mascot}
\alias{copy_phospho_mascot}
\title{Copy Mascot phospho \code{.csv}}
\usage{
copy_phospho_mascot(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Copy Mascot phospho \code{.csv}
}
