% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_phospho_maxquant}
\alias{copy_phospho_maxquant}
\title{Copy MaxQuant phospho \code{.txt}}
\usage{
copy_phospho_maxquant(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Copy MaxQuant phospho \code{.txt}
}
