% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_proteom_gtmt}
\alias{copy_proteom_gtmt}
\title{Copy proteoM global TMT \code{.txt}}
\usage{
copy_proteom_gtmt(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy proteoM global TMT \code{.txt}
}
