% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_proteom_plfq_2}
\alias{copy_proteom_plfq_2}
\title{Copy proteoM phospho LFQ \code{.txt}}
\usage{
copy_proteom_plfq_2(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy proteoM phospho LFQ \code{.txt}
}
