% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_specmill_plfq}
\alias{copy_specmill_plfq}
\title{Copy Spectrum Mill phospho LFQ \code{.ssv}}
\usage{
copy_specmill_plfq(dat_dir = "~/proteoQ/examples")
}
\arguments{
\item{dat_dir}{A character string to the working directory.}
}
\description{
Copy Spectrum Mill phospho LFQ \code{.ssv}
}
