% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_ssv}
\alias{copy_ssv}
\title{Copy Spectrum Mill \code{.ssv} files}
\usage{
copy_ssv(dat_dir, filelist)
}
\arguments{
\item{dat_dir}{A character string to the working directory.}

\item{filelist}{A list of files}
}
\description{
\code{copy_ssv} copies the Spectrum Mill outputs of \code{.ssv} files to a
target directory.
}
