% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_tsv}
\alias{copy_tsv}
\title{Copy MSFragger \code{.tsv} files}
\usage{
copy_tsv(dat_dir, filelist)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{filelist}{A list of files}
}
\description{
\code{copy_tsv} copies the MSFragger \code{psm.tsv} files to a target
directory.
}
