% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_txt}
\alias{copy_txt}
\title{Copy MaxQuant \code{.txt} files}
\usage{
copy_txt(dat_dir, filelist)
}
\description{
\code{copy_txt} copies the MaxQuant \code{msms.txt} files to a target
directory.
}
