% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_uniprot_hs}
\alias{copy_uniprot_hs}
\title{Copy \code{uniprot_hs_2014_07.fasta} to \code{db_path}}
\usage{
copy_uniprot_hs(db_path = "~\\\\proteoQ\\\\dbs\\\\fasta\\\\uniprot")
}
\arguments{
\item{db_path}{Character string; the local path for database(s).}
}
\description{
Copy \code{uniprot_hs_2014_07.fasta} to \code{db_path}
}
