% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_uniprot_hsmm}
\alias{copy_uniprot_hsmm}
\title{Copy \code{uniprot_hsmm_2020_03.fasta} to \code{db_path}}
\usage{
copy_uniprot_hsmm(db_path = "~\\\\proteoQ\\\\dbs\\\\fasta\\\\uniprot")
}
\arguments{
\item{db_path}{Character string; the local path for database(s).}
}
\description{
Copy \code{uniprot_hsmm_2020_03.fasta} to \code{db_path}
}
