% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_w2ref_exptsmry}
\alias{copy_w2ref_exptsmry}
\title{Copy an \code{expt_smry...} file to \code{dat_dir}}
\usage{
copy_w2ref_exptsmry(dat_dir)
}
\description{
Copy an \code{expt_smry...} file to \code{dat_dir}
}
