% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{cptac_expt_ref_w2}
\alias{cptac_expt_ref_w2}
\title{Copy an \code{expt_smry...} file to \code{dat_dir}}
\usage{
cptac_expt_ref_w2(dat_dir)
}
\description{
Copy an \code{expt_smry...} file to \code{dat_dir}
}
