% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{cptac_mq_expt_1}
\alias{cptac_mq_expt_1}
\title{Copy an \code{expt_smry...} file to \code{dat_dir}}
\usage{
cptac_mq_expt_1(dat_dir)
}
\description{
Copy an \code{expt_smry...} file to \code{dat_dir}
}
