% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{cptac_mqpep_expt}
\alias{cptac_mqpep_expt}
\title{Copy an \code{expt_smry...} file to \code{dat_dir}}
\usage{
cptac_mqpep_expt(dat_dir)
}
\description{
Copy an \code{expt_smry...} file to \code{dat_dir}
}
