% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003590}
\alias{F003590}
\title{CPTAC global data}
\format{A PSM table with header at a length of 39537}
\usage{
F003590
}
\description{
A \code{Mascot} PSM table from the \code{BI_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
