% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003593}
\alias{F003593}
\title{CPTAC global data}
\format{A PSM table with header at a length of 37240}
\usage{
F003593
}
\description{
A \code{Mascot} PSM table from the \code{JHU_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
