% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003594}
\alias{F003594}
\title{Mascot global TMT}
\format{
A PSM table with header at a length of 37951
}
\usage{
F003594
}
\description{
A \code{Mascot} PSM table from the \code{JHU_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
