% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003595}
\alias{F003595}
\title{Mascot global TMT}
\format{
A PSM table with header at a length of 41550
}
\usage{
F003595
}
\description{
A \code{Mascot} PSM table from the \code{PNNL_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
