% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003597}
\alias{F003597}
\title{CPTAC global data}
\format{A PSM table with header at a length of 42333}
\usage{
F003597
}
\description{
A \code{Mascot} PSM table from the \code{PNNL_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
