% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003598}
\alias{F003598}
\title{Mascot phosphopeptide TMT
A \code{Mascot} PSM table from the \code{BI_1} dataset. The PSM entries were
sampled at 10% from the full list.}
\format{
A PSM table with header at a length of 12761
}
\usage{
F003598
}
\description{
Mascot phosphopeptide TMT
A \code{Mascot} PSM table from the \code{BI_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
