% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003602}
\alias{F003602}
\title{Mascot phosphopeptide TMT}
\format{
A PSM table with header at a length of 12061
}
\usage{
F003602
}
\description{
A \code{Mascot} PSM table from the \code{BI_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
