% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003603}
\alias{F003603}
\title{Mascot phosphopeptide TMT}
\format{
A PSM table with header at a length of 9801
}
\usage{
F003603
}
\description{
A \code{Mascot} PSM table from the \code{JHU_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
