% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003604}
\alias{F003604}
\title{CPTAC phosphopeptide data}
\format{
A PSM table with header at a length of 10756
}
\usage{
F003604
}
\description{
A \code{Mascot} PSM table from the \code{JHU_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
