% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003605}
\alias{F003605}
\title{CPTAC phosphopeptide data}
\format{A PSM table with header at a length of 13089}
\usage{
F003605
}
\description{
A \code{Mascot} PSM table from the \code{PNNL_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
