% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003606}
\alias{F003606}
\title{CPTAC phosphopeptide data}
\format{A PSM table with header at a length of 11193}
\usage{
F003606
}
\description{
A \code{Mascot} PSM table from the \code{PNNL_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
