% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003795}
\alias{F003795}
\title{CPTAC global data}
\format{
A PSM table with header at a length of 188204
}
\usage{
F003795
}
\description{
A \code{Mascot} PSM table from the 10th and 15th fractions of \code{BI_1,
BI_2, JHU_1, JHU_2, PNNL_1, PNN_2} data. The data were searched against a
"Refseq Hu and Mm" database.
}
\keyword{datasets}
