% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{F003797}
\alias{F003797}
\title{CPTAC global data}
\format{
A PSM table with header at a length of 137479
}
\usage{
F003797
}
\description{
A \code{Mascot} PSM table from the 10th and 15th fractions of \code{BI_1,
BI_2, JHU_1, JHU_2, PNNL_1, PNN_2} data. The data were searched against a
"Refseq Hu only" database.
}
\keyword{datasets}
