% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{foo_fasta_rda}
\alias{foo_fasta_rda}
\title{Save \code{fasta} files to \code{.rda}}
\usage{
foo_fasta_rda(
  db_path = "~\\\\proteoQ\\\\dbs\\\\fasta\\\\refseq",
  from = "refseq_hs_2013_07.fasta"
)
}
\arguments{
\item{db_path}{Character string; the local path for database(s).}

\item{from}{The source \code{.xlsx} file name.}
}
\description{
Save \code{fasta} files to \code{.rda}
}
\examples{
\donttest{
foo_fasta_rda("~\\\\proteoQ\\\\dbs\\\\fasta\\\\uniprot", "uniprot_hs_2014_07.fasta")
foo_fasta_rda("~\\\\proteoQ\\\\dbs\\\\fasta\\\\refseq", "refseq_hs_2013_07.fasta")
}
}
