% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{foo_mgf}
\alias{foo_mgf}
\title{Saves MGFs from MSConvert or Proteome Discoverer to \code{.rda}.}
\usage{
foo_mgf(dat_dir = "~/proteoQ/examples", filelist = c("pd_bi1_gl_partial"))
}
\arguments{
\item{dat_dir}{A character string to the working directory.}

\item{filelist}{A list of MGF files.}
}
\description{
Saves MGFs from MSConvert or Proteome Discoverer to \code{.rda}.
}
\examples{
\donttest{
foo_mgf(filelist = c("pd_bi1_gl_partial", "msconv_bi1_imac_partial"))
}
}
