% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{msconv_bi1_imac_partial}
\alias{msconv_bi1_imac_partial}
\title{MSConvert IMAC TMT}
\format{
A MGF file at a length of 800889 lines.
}
\usage{
msconv_bi1_imac_partial
}
\description{
A small subset of MGF entries from the \code{BI_1} for demonstration with
\href{https://github.com/qzhang503/proteoM}{proteoM}. The MGF generator is
MSConvert.
}
\keyword{datasets}
