% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{msms}
\alias{msms}
\title{CPTAC LFQ data from MaxQuant}
\format{
A PSM table at a length of 139561.
}
\usage{
msms
}
\description{
A \code{MaxQuant} LFQ PSM table from the 10th and 15th fractions of
\code{BI_1, BI_2, JHU_1, JHU_2, PNNL_1, PNN_2} data. The data were searched
against a "Refseq Hu and Mm" database. Some columns were excluded to reduce
the size of the file.
}
\keyword{datasets}
