% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{msms_bi_p1}
\alias{msms_bi_p1}
\title{CPTAC phosphopeptide data}
\format{
A PSM table with header at a length of 10711
}
\usage{
msms_bi_p1
}
\description{
A \code{MaxQuant} PSM table from the \code{BI_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
