% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{msms_bi_p2}
\alias{msms_bi_p2}
\title{CPTAC phosphopeptide data}
\format{A PSM table with header at a length of 10031}
\usage{
msms_bi_p2
}
\description{
A \code{MaxQuant} PSM table from the \code{BI_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
