% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{msms_jhu_1}
\alias{msms_jhu_1}
\title{MaxQuant global TMT}
\format{
A PSM table with header at a length of 25552
}
\usage{
msms_jhu_1
}
\description{
A \code{MaxQuant} PSM table from the \code{JHU_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
