% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{msms_pnnl_2}
\alias{msms_pnnl_2}
\title{CPTAC global data}
\format{A PSM table with header at a length of 28691}
\usage{
msms_pnnl_2
}
\description{
A \code{MaxQuant} PSM table from the \code{PNNL_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
