% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{msms_pnnl_p1}
\alias{msms_pnnl_p1}
\title{MaxQuant phosphopeptide TMT}
\format{
A PSM table with header at a length of 11117
}
\usage{
msms_pnnl_p1
}
\description{
A \code{MaxQuant} PSM table from the \code{PNNL_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
