% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{msms_pnnl_p2}
\alias{msms_pnnl_p2}
\title{MaxQuant phosphopeptide TMT}
\format{
A PSM table with header at a length of 9926
}
\usage{
msms_pnnl_p2
}
\description{
A \code{MaxQuant} PSM table from the \code{PNNL_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
