% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pd_bi1_gl_partial}
\alias{pd_bi1_gl_partial}
\title{Proteome Discoverer (PD) global TMT}
\format{
A MGF file at a length of 999708 entries.
}
\usage{
pd_bi1_gl_partial
}
\description{
A small subset of MGF entries from the \code{BI_1}. The MGF generator is PD.
}
\keyword{datasets}
