% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{peptides}
\alias{peptides}
\title{CPTAC LFQ data from MaxQuant}
\format{
A peptide table at a length of 57517.
}
\usage{
peptides
}
\description{
A \code{MaxQuant} LFQ peptide table corresponding to data set "msms_lfq".
}
\keyword{datasets}
