% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psm_bi_1}
\alias{psm_bi_1}
\title{CPTAC global data}
\format{
A PSM table of 31480 rows and 32 columns
}
\usage{
psm_bi_1
}
\description{
A \code{MSFragger} LFQ PSM table from the \code{BI_1} dataset. The PSM entries were
from offline fractions 10 and 15.
}
\keyword{datasets}
