% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psm_lfq_bi_p1}
\alias{psm_lfq_bi_p1}
\title{MSFragger IMAC data (LFQ)}
\format{
A PSM table of 24403 rows and 32 columns
}
\usage{
psm_lfq_bi_p1
}
\description{
A \code{MSFragger} LFQ PSM table from the \code{BI_1} dataset. The PSM entries were
from offline fractions 5 and 10 and searched against UniProt.
}
\keyword{datasets}
