% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psm_lfq_pnnl_p2}
\alias{psm_lfq_pnnl_p2}
\title{MSFragger IMAC data (LFQ)}
\format{
A PSM table of 24197 rows and 32 columns
}
\usage{
psm_lfq_pnnl_p2
}
\description{
A \code{MSFragger} LFQ PSM table from the \code{PNNL_2} dataset. The PSM entries were
from offline fractions 5 and 10 and searched against UniProt.
}
\keyword{datasets}
