% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psm_tmt_bi_2}
\alias{psm_tmt_bi_2}
\title{CPTAC global data}
\format{
A PSM table of 35854 rows and 44 columns
}
\usage{
psm_tmt_bi_2
}
\description{
A \code{MSFragger} TMT PSM table from the \code{BI_2} dataset. The PSM entries were
from offline fractions 10 and 15.
}
\keyword{datasets}
