% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psm_tmt_jhu_1}
\alias{psm_tmt_jhu_1}
\title{MSFragger global TMT}
\format{
A PSM table of 35657 rows and 44 columns
}
\usage{
psm_tmt_jhu_1
}
\description{
A \code{MSFragger} TMT PSM table from the \code{JHU_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
