% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psm_tmt_pnnl_2}
\alias{psm_tmt_pnnl_2}
\title{MSFragger global TMT}
\format{
A PSM table of 40018 rows and 44 columns
}
\usage{
psm_tmt_pnnl_2
}
\description{
A \code{MSFragger} TMT PSM table from the \code{PNNL_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
