% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PSMexport_bi_p1}
\alias{PSMexport_bi_p1}
\title{CPTAC phosphopeptide data}
\format{A PSM table with header at a length of 9007}
\usage{
PSMexport_bi_p1
}
\description{
A \code{Spectrum Mill} PSM table from the \code{BI_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
