% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PSMexport_jhu_1}
\alias{PSMexport_jhu_1}
\title{CPTAC global data}
\format{A PSM table with header at a length of 28480}
\usage{
PSMexport_jhu_1
}
\description{
A \code{Spectrum Mill} PSM table from the \code{JHU_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
