% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PSMexport_jhu_2}
\alias{PSMexport_jhu_2}
\title{CPTAC global data}
\format{A PSM table with header at a length of 28766}
\usage{
PSMexport_jhu_2
}
\description{
A \code{Spectrum Mill} PSM table from the \code{JHU_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
