% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PSMexport_jhu_p2}
\alias{PSMexport_jhu_p2}
\title{SM phosphopeptide TMT}
\format{
A PSM table with header at a length of 7385
}
\usage{
PSMexport_jhu_p2
}
\description{
A \code{Spectrum Mill} PSM table from the \code{JHU_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
