% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PSMexport_pnnl_2}
\alias{PSMexport_pnnl_2}
\title{SM global TMT}
\format{
A PSM table with header at a length of 28999
}
\usage{
PSMexport_pnnl_2
}
\description{
A \code{Spectrum Mill} PSM table from the \code{PNNL_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
