% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PSMexport_pnnl_p1}
\alias{PSMexport_pnnl_p1}
\title{CPTAC phosphopeptide data}
\format{A PSM table with header at a length of 8045}
\usage{
PSMexport_pnnl_p1
}
\description{
A \code{Spectrum Mill} PSM table from the \code{PNNL_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
