% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psmQ_lfq_p}
\alias{psmQ_lfq_p}
\title{proteoM IMAC data (LFQ)}
\format{
A PSM table at a length of 308402
}
\usage{
psmQ_lfq_p
}
\description{
A \code{proteoM} LFQ PSM table from the 5th and 10th fractions of
\code{BI_1, BI_2, JHU_1, JHU_2, PNNL_1, PNN_2} IMAC. The data were searched
against a "UniProt Hu and Mm" database.
}
\details{
Redundancy at neutral losses and variable modifications kept
}
\keyword{datasets}
