% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psmQ_lfq_pnnl_p1}
\alias{psmQ_lfq_pnnl_p1}
\title{proteoM IMAC data (LFQ)}
\format{
A PSM table of 17057 rows and 46 columns
}
\usage{
psmQ_lfq_pnnl_p1
}
\description{
A \code{proteoM} LFQ PSM table from the \code{PNNL_1} dataset. The PSM entries
were from offline fractions 5 and 10 and searched against UniProt.
}
\keyword{datasets}
