% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psmQ_tmt_jhu_2}
\alias{psmQ_tmt_jhu_2}
\title{proteoM global TMT}
\format{
A PSM table with header at a length of 57057
}
\usage{
psmQ_tmt_jhu_2
}
\description{
A \code{proteoM} PSM table from the \code{JHU_2} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
