% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psmQ_tmt_pnnl_1}
\alias{psmQ_tmt_pnnl_1}
\title{proteoM global TMT}
\format{
A PSM table with header at a length of 77035
}
\usage{
psmQ_tmt_pnnl_1
}
\description{
A \code{proteoM} PSM table from the \code{PNNL_1} dataset. The PSM entries were
sampled at 10% from the full list.
}
\keyword{datasets}
