% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{scc_hs}
\alias{scc_hs}
\title{UniProt subcellular locations}
\format{
A table of columns "uniprot_acc" and "scc_1", "scc_2" etc. at a
  length of 16697.
}
\usage{
scc_hs
}
\description{
A table complied from UniProt with columns "Entry" and "Subcellular location [CC]".
}
\keyword{datasets}
