% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{scc_mm}
\alias{scc_mm}
\title{UniProt subcellular locations}
\format{
A table of columns "uniprot_acc" and "scc_1", "scc_2" etc. at a
  length of 14540.
}
\usage{
scc_mm
}
\description{
A table compiled from UniProt with columns "Entry" and "Subcellular location [CC]".
}
\keyword{datasets}
