% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_csv}
\alias{copy_csv}
\title{Copy Mascot \code{.csv} files}
\usage{
copy_csv(dat_dir, filelist, mapped = NULL)
}
\arguments{
\item{dat_dir}{A character string to the working directory.}

\item{filelist}{A list of files}
}
\description{
\code{copy_csv} copies the Mascot outputs of \code{.csv} files to a target
directory.
}
