% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_csv}
\alias{copy_csv}
\title{Copy Mascot \code{.csv} files}
\usage{
copy_csv(dat_dir, filelist)
}
\description{
\code{copy_dat} copies the Mascot outputs of \code{.csv} files to a target
directory.
}
