% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_fasta}
\alias{copy_fasta}
\title{Copy \code{fasta} files}
\usage{
copy_fasta(
  db_path = "~/proteoQ/dbs/fasta/uniprot",
  file = "uniprot_hsmm_2020_03.rda"
)
}
\arguments{
\item{db_path}{Character string; the local path for database(s).}

\item{file}{Character string; the file name of a target \code{.rda}.}
}
\description{
\code{copy_fasta} copies \code{fasta} files to a target directory.
}
