% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_refseq_hs}
\alias{copy_refseq_hs}
\title{Copy \code{refseq_hs_2013_07.fasta}}
\usage{
copy_refseq_hs(fasta_dir = "~\\\\proteoQ\\\\dbs\\\\refseq")
}
\description{
Copy \code{refseq_hs_2013_07.fasta}
}
