% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_refseq_mm}
\alias{copy_refseq_mm}
\title{Copy \code{refseq_mm_2013_07.fasta}}
\usage{
copy_refseq_mm(fasta_dir = "~\\\\proteoQ\\\\dbs\\\\fasta\\\\refseq")
}
\description{
Copy \code{refseq_mm_2013_07.fasta}
}
