% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_txt}
\alias{copy_txt}
\title{Copy MaxQuant \code{.txt} files}
\usage{
copy_txt(dat_dir, filelist)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}

\item{filelist}{A list of files}
}
\description{
\code{copy_txt} copies the MaxQuant \code{msms.txt} files to a target
directory.
}
