% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_w2ref_exptsmry}
\alias{copy_w2ref_exptsmry}
\title{Copy a metadata file \code{expt_smry.xlsx} to \code{dat_dir}}
\usage{
copy_w2ref_exptsmry(dat_dir)
}
\arguments{
\item{dat_dir}{A character string to the working directory. The default is to
match the value under the global environment.}
}
\description{
Copy a metadata file \code{expt_smry.xlsx} to \code{dat_dir}
}
